fix: don't overwrite existing tracks when downloading
It seems to be possible for two tracks to have the same identifier, for
example:
spotify:track:5bGy2uC2zUDmveSkq6Fnoo
spotify:track:7grZDGr08mjdXeL03zTseu
are actually the same song but one is from a single and the other one is
from the album. Even tough spotify has a method for relating two tracks
using the 'alternatives' field in the track metadata this does not seem
to be present in all tracks.

Right now the solution is to just rename the new file adding the id at
the end of the filename, in the future something better could done like
checking the hash of the files. At first look both of the tracks had the
exact same duration so it is likely their hash will be the same.

There is also a concept of a `FileId` in spotify. It is possible that
these two songs have the same FileId and therefore we should ignore
based on that.