fix: remove dots from filenames when organizing
Using dots in a filename caused problems when setting the extension of
the file. It seems that the default method in `PathBuf` just replaces
everything starting at the first dot. This might no be an ideal fix but
since the file metadata still includes the correct name it shouldn't be
a problem.