implemented workaround for failing lyrics downloads
For some reason there started to be failures when downloading tracks.
The errors have the following message:
	Client specified an invalid argument { Response status code: 400 Bad Request }
I tracked this down to the lyrics download and I don't since this is being made
by the spotify library that I don't controll or know much about the workaround
implemented will instead just return no lyrics and print a warning instead of
failing to fetch metadata from the track.